% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_level.R
\name{geoess_block_spectrum}
\alias{geoess_block_spectrum}
\title{Block ESS Spectrum}
\usage{
geoess_block_spectrum(Iunit, Ipi)
}
\arguments{
\item{Iunit}{Unit Fisher information for a block.}

\item{Ipi}{Prior curvature for a block.}
}
\value{
A list with eigenvalues and eigenvectors.
}
\description{
Block ESS Spectrum
}
