% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_core.R
\name{geoess_dir}
\alias{geoess_dir}
\title{Directional ESS (Alias)}
\usage{
geoess_dir(I1, Ipi, v)
}
\arguments{
\item{I1}{Numeric matrix. Fisher information per observation.}

\item{Ipi}{Numeric matrix. Prior curvature.}

\item{v}{Numeric vector direction.}
}
\value{
Numeric scalar ESS for direction v.
}
\description{
Directional ESS (Alias)
}
