% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_core.R
\name{geoess_eig}
\alias{geoess_eig}
\title{Eigen ESS Spectrum (Alias)}
\usage{
geoess_eig(I1, Ipi, symmetrize = TRUE)
}
\arguments{
\item{I1}{Numeric matrix. Fisher information per observation.}

\item{Ipi}{Numeric matrix. Prior curvature.}

\item{symmetrize}{Logical. If TRUE, symmetrize inputs before eigen calculations.}
}
\value{
Numeric vector of eigenvalues.
}
\description{
Eigen ESS Spectrum (Alias)
}
