% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_ess.R
\name{geoess_global_design}
\alias{geoess_global_design}
\title{Design-Based Global ESS}
\usage{
geoess_global_design(rQ, I1, Ipi, ...)
}
\arguments{
\item{rQ}{Function \code{rQ(n)} returning draws from a design distribution.}

\item{I1}{Fisher information per observation.}

\item{Ipi}{Prior curvature.}

\item{...}{Passed to \code{geoess_global}.}
}
\value{
A \code{geoess_global_result}.
}
\description{
Design-Based Global ESS
}
