% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_level.R
\name{geoess_hier}
\alias{geoess_hier}
\title{Hierarchical ESS Dispatcher}
\usage{
geoess_hier(method = c("level", "between_borrowed"), ...)
}
\arguments{
\item{method}{Character. "level" or "between_borrowed".}

\item{...}{Passed to the selected method.}
}
\value{
Hierarchical ESS result.
}
\description{
Hierarchical ESS Dispatcher
}
