% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier.R
\name{geoess_Ipi_hier}
\alias{geoess_Ipi_hier}
\title{Hierarchical Curvature via Schur Complement}
\usage{
geoess_Ipi_hier(
  Htt,
  Htp,
  Hpp,
  symmetrize = TRUE,
  solver = c("chol", "solve"),
  return_components = TRUE
)
}
\arguments{
\item{Htt}{Numeric matrix. Curvature block for \eqn{\theta}.}

\item{Htp}{Numeric matrix. Cross curvature block.}

\item{Hpp}{Numeric matrix. Curvature block for \eqn{\phi}.}

\item{symmetrize}{Logical; if TRUE, symmetrize outputs.}

\item{solver}{Character. "chol" or "solve".}

\item{return_components}{Logical; if TRUE, return all blocks.}
}
\value{
A list with effective/between/borrowed curvature blocks.
}
\description{
Hierarchical Curvature via Schur Complement
}
