% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier.R
\name{geoess_Ipi_hier_from_fn}
\alias{geoess_Ipi_hier_from_fn}
\title{Hierarchical Curvature from Functions}
\usage{
geoess_Ipi_hier_from_fn(
  theta_ref,
  phi_ref,
  Htt_fn = NULL,
  Htp_fn = NULL,
  Hpp_fn = NULL,
  logprior_joint = NULL,
  ...
)
}
\arguments{
\item{theta_ref}{Numeric vector for \eqn{\theta^*}.}

\item{phi_ref}{Numeric vector for \eqn{\phi^*}.}

\item{Htt_fn}{Optional function returning Htt.}

\item{Htp_fn}{Optional function returning Htp.}

\item{Hpp_fn}{Optional function returning Hpp.}

\item{logprior_joint}{Optional joint log-prior function.}

\item{...}{Passed to user functions.}
}
\value{
A list with curvature blocks and effective components.
}
\description{
Hierarchical Curvature from Functions
}
