% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier.R
\name{geoess_Ipi_hier_from_mix}
\alias{geoess_Ipi_hier_from_mix}
\title{Hierarchical Curvature from Joint Mixture}
\usage{
geoess_Ipi_hier_from_mix(
  mix_joint,
  theta_ref,
  phi_ref,
  psd = c("none", "clip", "nearPD"),
  ...
)
}
\arguments{
\item{mix_joint}{A \code{geoess_mix} object on joint (theta, phi).}

\item{theta_ref}{Numeric vector for \eqn{\theta^*}.}

\item{phi_ref}{Numeric vector for \eqn{\phi^*}.}

\item{psd}{Character. Passed to \code{geoess_Ipi_from_mix}.}

\item{...}{Passed to mixture calculus.}
}
\value{
A list with curvature blocks and effective components.
}
\description{
Hierarchical Curvature from Joint Mixture
}
