% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_mvnorm_calculus.R
\name{geoess_mix_gradlog}
\alias{geoess_mix_gradlog}
\title{MVN Mixture Gradient of Log-Density}
\usage{
geoess_mix_gradlog(mix, theta, use_cache = TRUE)
}
\arguments{
\item{mix}{A \code{geoess_mix} object with MVN components.}

\item{theta}{Numeric vector evaluation point.}

\item{use_cache}{Logical; if TRUE, uses cached precision/Cholesky if present.}
}
\value{
Numeric vector gradient.
}
\description{
MVN Mixture Gradient of Log-Density
}
