% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_mvnorm_calculus.R
\name{geoess_mix_hesslog}
\alias{geoess_mix_hesslog}
\title{MVN Mixture Hessian of Log-Density}
\usage{
geoess_mix_hesslog(mix, theta, use_cache = TRUE, symmetrize = TRUE)
}
\arguments{
\item{mix}{A \code{geoess_mix} object with MVN components.}

\item{theta}{Numeric vector evaluation point.}

\item{use_cache}{Logical; if TRUE, uses cached precision/Cholesky if present.}

\item{symmetrize}{Logical; if TRUE, symmetrize the result.}
}
\value{
Numeric matrix Hessian of log q(theta).
}
\description{
MVN Mixture Hessian of Log-Density
}
