% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_normal.R
\name{geoess_normal_mu}
\alias{geoess_normal_mu}
\title{Geometric ESS for Normal Mean with Known Variance}
\usage{
geoess_normal_mu(sigma, s0)
}
\arguments{
\item{sigma}{Numeric. Data standard deviation.}

\item{s0}{Numeric. Prior standard deviation.}
}
\value{
Numeric scalar ESS, \eqn{\sigma^2/s_0^2}.
}
\description{
Geometric ESS for Normal Mean with Known Variance
}
\examples{
geoess_normal_mu(sigma = 2, s0 = 4)
}
