% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_core.R
\name{geoess_trace}
\alias{geoess_trace}
\title{Trace ESS (Scalar)}
\usage{
geoess_trace(I1, Ipi, symmetrize = TRUE)
}
\arguments{
\item{I1}{Numeric matrix. Fisher information per observation.}

\item{Ipi}{Numeric matrix. Prior curvature.}

\item{symmetrize}{Logical. If TRUE, symmetrize inputs before calculations.}
}
\value{
Numeric scalar ESS.
}
\description{
Trace ESS (Scalar)
}
