% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_normal.R
\name{I1_normal_mu}
\alias{I1_normal_mu}
\title{Fisher Information for Normal Mean (Per Observation)}
\usage{
I1_normal_mu(sigma)
}
\arguments{
\item{sigma}{Numeric. Data standard deviation.}
}
\value{
1x1 matrix for \eqn{1/\sigma^2}.
}
\description{
Fisher Information for Normal Mean (Per Observation)
}
\examples{
I1_normal_mu(2)
}
