% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_normal.R
\name{Ipi_normal_mu}
\alias{Ipi_normal_mu}
\title{Prior Curvature for Normal Mean (Normal Prior)}
\usage{
Ipi_normal_mu(s0)
}
\arguments{
\item{s0}{Numeric. Prior standard deviation.}
}
\value{
1x1 matrix for \eqn{1/s_0^2}.
}
\description{
Prior Curvature for Normal Mean (Normal Prior)
}
\examples{
Ipi_normal_mu(4)
}
