% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_beta.R
\name{mixbeta}
\alias{mixbeta}
\title{Beta Mixture Distributions}
\usage{
mixbeta(..., param = c("ab", "ms", "mn"))
}
\arguments{
\item{...}{Components as numeric vectors of length 3 or a 3xK matrix.
Each component is \code{c(weight, p1, p2)} where \code{p1, p2} are
parameterized by \code{param}.}

\item{param}{Character. "ab" for shape1/shape2, "ms" for mean/sd,
or "mn" for mean/effective sample size.}
}
\value{
A matrix of class \code{c("betaMix", "mix")} with rows
  \code{w}, \code{a}, \code{b}.
}
\description{
Create and evaluate finite mixtures of Beta distributions.
}
