% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_ess_result.R
\name{print.geoess_global_result}
\alias{print.geoess_global_result}
\title{Global ESS Result Methods}
\usage{
\method{print}{geoess_global_result}(x, ...)
}
\arguments{
\item{x}{A \code{geoess_global_result} object.}

\item{...}{Unused.}
}
\value{
The object, invisibly for \code{print}.
}
\description{
Global ESS Result Methods
}
